added 'arm64-v8a' to abi filters, changed gsize&gssize types to use long ints, upgraded to gradle to 8.3.0, added some warning suppression flags, set min SDK version to 19, updated revision to 1.1.4a

+++ app/src/main/jni/glib/android/glibconfig.h
@@ -56,8 +56,8 @@ G_GNUC_EXTENSION typedef unsigned long long guint64;
 #define GLIB_SIZEOF_LONG   4
 #define GLIB_SIZEOF_SIZE_T 4
 
-typedef signed int gssize;
-typedef unsigned int gsize;
+typedef signed long int gssize;
+typedef unsigned long int gsize;
 #define G_GSIZE_MODIFIER ""
 #define G_GSSIZE_FORMAT "i"
 #define G_GSIZE_FORMAT "u"
